requirements.txt
Django==3.2.7
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
drf-yasg==1.21.10
gunicorn==23.0.0
whitenoise==6.9.0
psycopg2
django-filter